style: 스타일 속성 변경
