+++ src/page/home/Home.tsx
@@ -16,25 +16,6 @@ import { useLoginModal } from '@/common/hook/useLoginModal';
 const sectionKeys = ['mandalart', 'ai', 'todo'] as const;
 
 const Home = () => {
-  useEffect(() => {
-    const getToken = async () => {
-      const code = getGoogleAuthCode();
-
-      if (!code) {
-        return;
-      }
-
-      try {
-        const data = await getAccessToken(code);
-        console.log('로그인 성공:', data);
-      } catch (error) {
-        console.error('로그인 실패:', error);
-      }
-    };
-
-    getToken();
-  }, []);
-
   const scrolls = useMultipleFadeInOnView();
   const end = useFadeInOnView<HTMLDivElement>();
