+++ src/page/signup/SignUp.tsx
@@ -7,6 +7,7 @@ import SignUpButton from '@/page/signup/component/SignUpButton/SignUpButton';
 import * as styles from '@/page/signup/SignUp.css';
 import { useSignUpForm } from '@/page/signup/hook/useSignUpForm';
 import { PATH } from '@/route';
+import { useGoogleAuthEffect } from '@/page/signup/hook/useGoogleAuthEffect';
 
 const SIGNUP_MESSAGE = '회원가입 후 NiNE DOT를 만나보세요!';
 const FIT_INFO_MESSAGE = '내 성향을 선택하고 맞춤형 목표 추천을 받아보세요';
@@ -25,6 +26,9 @@ const SignUp = () => {
   const handleSignUp = () => {
     navigate(PATH.INTRO);
   };
+
+  const data = useGoogleAuthEffect();
+  console.log(data);
 
   return (
     <main className={styles.mainContainer}>
